% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CumVec.R
\name{CumVec}
\alias{CumVec}
\title{CumVec This function allow you to add a vector in a squence to get cumulative values}
\usage{
CumVec(x)
}
\arguments{
\item{x}{a vector to increment}
}
\value{
a vector of incremented values in a sequence relative to the paretn vector
}
\description{
CumVec This function allow you to add a vector in a squence to get cumulative values
}
