% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likert_scale_conversion.R
\name{likert_scale_conversion}
\alias{likert_scale_conversion}
\title{Correlation exploration using vectors or data frames}
\usage{
likert_scale_conversion(df)
}
\arguments{
\item{df}{data frame or a matrix with likert scale rating strongly agree, agree, etc}
}
\value{
a matrix of plots
}
\description{
likert_scale_conversion
convert likert scale characters to numerical numbers
It takes on an array or a data frame. if a vector is passed to the function it will be converted into a data frame
Check if all the variables are numeric and not strings
}
\author{
Richard Magala, 2022
}
