% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_corr.R
\name{mat_corr}
\alias{mat_corr}
\title{mat_corr
generate matrix table of significant variables}
\usage{
mat_corr(y, df, showSig = F, sig.level = 0.05)
}
\arguments{
\item{y}{is a response variable for exploring its relationship with the other predictor variables}

\item{df}{data frame or a matrix with predictor variables}

\item{showSig}{returns names of the significant variables at the specifies significant level}

\item{sig.level}{specifies the significant level for deciding a significant relation between variables default is 0.05.}
}
\value{
A data frame of correlation, p-values and their t-values
}
\description{
This function explores the correlation of a variables with multiple dependent variables.
It takes on an array or a data frame. if a vector is passed to the function it will be converted into a data frame
Check if all the variables are numeric and not strings
}
