% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_regression.R
\name{mult_regression}
\alias{mult_regression}
\title{mult_regression
run multiple regresson on a data frame}
\usage{
mult_regression(y, df)
}
\value{
modelcountainer_summaries
}
\description{
mult_regression
run multiple regresson on a data frame
}
\author{
richard magala
}
