% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matr.R
\name{plot_matr}
\alias{plot_matr}
\title{Correlation exploration using vectors or data frames}
\usage{
plot_matr(y, df, nrow = 2, ncol = 2, sig.level = 0.05, ylab = "")
}
\arguments{
\item{y}{is a response variable for exploring its relationship with the other predictor variables}

\item{df}{data frame or a matrix with predictor variables}

\item{nrow}{specifies the number of plots to be inserted in a row}

\item{ncol}{specifies the number of plots to be inserted in a column}

\item{sig.level}{specifies the number significant level default is 0.05}

\item{ylab}{specifies the y-axis, which is the name of the respons variables defaults is the y names}
}
\value{
a matrix of plots
}
\description{
plot_matr
generate plots and fitted lines of significantly correlated variables for exploration.
}
\details{
This function allows you to visualize the correlated variables with the selected variable. the fitted line is through regression
It takes on an array or a data frame. if a vector is passed to the function it will be converted into a data frame
Check if all the variables are numeric and not strings
}
\author{
Richard Magala, 2022
}
